requirements.txt
selenium==4.38.0
python-dotenv
python==3.10.1